revised internal plot engine to use double precision for layout calculations
git-svn-id: https://efmsvn2.nrel.gov/wex/svn/trunk@449 e21c72fa-1056-4880-a922-3284d9a13290

+++ include/wex/plot/plhistplot.h
@@ -17,7 +17,7 @@ public:
 	void SetData( const std::vector<wxRealPoint> &data );
 
 	//Getters and Setters
-	void SetLineStyle( const wxColour &c, int width );
+	void SetLineStyle( const wxColour &c, double width );
 	void SetFillColour( const wxColour &c );
 	void SetNumberOfBins( size_t n );
 	void SetNormalize( NormalizeType n );
@@ -50,7 +50,7 @@ private:
 	bool m_ignoreZeros;
 
 	wxColour m_lineColour;
-	int m_lineThickness;
+	double m_lineThickness;
 	wxColour m_fillColour;
 
 	size_t m_numberOfBins;
